Detect Dupes
Detect dupes script allow you to give a folder path as input and a folder as output within the output folder will be a folder per file that is the same EG images , documents and videos as long as they have the exact same bytes the MD5 hash will be the same, very useful when trying to minimise how many images are stored on a server. Could make an alteration to the script to create symlinks between the items so you dont have to store multiple images only one image linked to multiple names.